Port permission api module to 1.6.4
