+++ legacy-fabric-permissions-api-v1/1.6.4/gradle.properties
@@ -1,0 +1,2 @@
+minVersionIncluded=1.3
+maxVersionIncluded=1.6.4
