+++ legacy-fabric-permissions-api-v1/1.6.4/src/main/resources/fabric.mod.json
