+++ legacy-fabric-permissions-api-v1/1.6.4/src/main/resources/legacy-fabric-permissions-api-v1.mixins.json
@@ -3,11 +3,7 @@
   "package": "net.legacyfabric.fabric.mixin.permission",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "CommandBlockExecutorMixin",
-    "CommandStats_1Mixin",
-    "ExecuteCommand_1Mixin",
-    "SignBlockEntity_1Mixin",
-    "SignBlockEntity_2Mixin"
+    "CommandBlockExecutorMixin"
   ],
   "injectors": {
     "defaultRequire": 1
