+++ legacy-fabric-permissions-api-v1/1.8.9/gradle.properties
@@ -1,2 +1,2 @@
-minVersionIncluded=1.3
+minVersionIncluded=1.7.10
 maxVersionIncluded=1.12.2
