+++ legacy-fabric-permissions-api-v1/common/src/main/java/net/legacyfabric/fabric/mixin/permission/ServerPlayerEntityMixin.java
